Move AST methods from lib/ast/api to the respective classes
No more prototype fiddling.
