+++ lib/ast/base.js
@@ -79,6 +79,13 @@ module.exports = class Node {
     }
 
     /* istanbul ignore next */
+    /**
+     * Optimize this AST node.
+     *
+     * Optimization removes redundant operations and converts ThingTalk to canonical form.
+     *
+     * @returns {Ast~Node} the optimized node
+     */
     optimize() {
         return this;
     }
