+++ lib/ast/class_def.js
@@ -15,7 +15,7 @@ const Type = require('../type');
 const { prettyprintClassDef } = require('../prettyprint');
 const { clean, cleanKind } = require('../utils');
 const { Value } = require('./values');
-const { InputParam } = require('./primitive');
+const { InputParam } = require('./expression');
 const { Statement } = require('./program');
 const { FunctionDef } = require('./function_def');
 const { getString, extractImports, typeToHTML } = require('./manifest_utils');
