+++ lib/ast/expression.js
@@ -11,8 +11,22 @@
 const assert = require('assert');
 
 const Node = require('./base');
-const { ExpressionSignature } = require('./function_def');
+const {
+    ArgumentDef,
+    ExpressionSignature
+} = require('./function_def');
 const { Value } = require('./values');
+
+const { prettyprintFilterExpression } = require('../prettyprint');
+const Typechecking = require('../typecheck');
+const Optimizer = require('../optimize');
+const {
+    iterateSlots2InputParams,
+    recursiveYieldArraySlots,
+    makeScope,
+    DeviceAttributeSlot,
+    FilterSlot,
+} = require('./slots');
 
 /**
  * Base class of all expressions that select a device.
@@ -253,6 +267,42 @@ class Invocation extends Node {
         const in_params = this.in_params && this.in_params.length > 0 ? this.in_params.toString() : '';
         return `Invocation(${this.selector.toString()}, ${this.channel}, ${in_params}, )`;
     }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this invocation.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.Invocation#iterateSlots2} instead.
+     */
+    *iterateSlots(scope) {
+        yield [null, this.selector, this, null];
+        for (let in_param of this.in_params)
+            yield [this.schema, in_param, this, scope];
+        return [this, makeScope(this)];
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this invocation.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2(scope) {
+        if (this.selector.isDevice) {
+            for (let attr of this.selector.attributes)
+                yield new DeviceAttributeSlot(this, attr);
+
+            // note that we yield the selector after the device attributes
+            // this way, almond-dialog-agent will first ask any question to slot-fill
+            // the device attributes (if somehow it needs to) and then use the chosen
+            // device attributes to choose the device
+            yield this.selector;
+        }
+        return yield* iterateSlots2InputParams(this, scope);
+    }
 }
 Invocation.prototype.isInvocation = true;
 module.exports.Invocation = Invocation;
@@ -277,7 +327,180 @@ module.exports.Invocation = Invocation;
  * @property {boolean} isVarRef - true if this is {@link Ast.BooleanExpression.VarRef}
  * @property {boolean} isCompute - true if this is {@link Ast.BooleanExpression.Compute}
  */
-class BooleanExpression extends Node {}
+class BooleanExpression extends Node {
+    /**
+     * Typecheck this boolean expression.
+     *
+     * This method can be used to typecheck a boolean expression is isolation,
+     * outside of a ThingTalk program.
+     *
+     * @param {Ast.ExpressionSignature} schema - the signature of the query expression this filter
+     *                                           would be attached to
+     * @param {null} scope - reserved, must be null
+     * @param {SchemaRetriever} schemas - schema retriever object to retrieve Thingpedia information
+     * @param {Object.<string,Ast.ClassDef>} classes - additional locally defined classes, overriding Thingpedia
+     * @param {boolean} [useMeta=false] - retreive natural language metadata during typecheck
+     * @alias Ast.BooleanExpression#typecheck
+     */
+    typecheck(schema, scope, schemas, classes, useMeta) {
+        return Typechecking.typeCheckFilter(this, schema, scope, schemas, classes, useMeta);
+    }
+
+    /**
+     * Convert this boolean expression to prettyprinted ThingTalk code.
+     *
+     * @param {string} [prefix] - prefix each output line with this string (for indentation)
+     * @return {string} the prettyprinted code
+     * @alias Ast.BooleanExpression#prettyprint
+     */
+    prettyprint() {
+        return prettyprintFilterExpression(this);
+    }
+
+    optimize() {
+        return Optimizer.optimizeFilter(this);
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this boolean expression.
+     *
+     * @alias Ast.BooleanExpression#iterateSlots
+     * @param {Ast.ExpressionSignature} schema - the signature of the query expression this filter is attached to
+     * @param {Ast.Invocation} prim - the nearest primitive
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.BooleanExpression#iterateSlots2} instead.
+     */
+    *iterateSlots(schema, prim, scope) {
+        if (this.isTrue || this.isFalse)
+            return;
+        if (this.isAnd || this.isOr) {
+            for (let op of this.operands)
+                yield* op.iterateSlots(schema, prim, scope);
+            return;
+        }
+        if (this.isNot) {
+            yield* this.expr.iterateSlots(schema, prim, scope);
+            return;
+        }
+        if (this.isExternal) {
+            yield* Invocation.prototype.iterateSlots.call(this, scope);
+            yield* this.filter.iterateSlots(this.schema, prim, makeScope(this));
+            return;
+        }
+        if (this.isVarRef) {
+            // TODO
+            return;
+        }
+        if (this.isCompute) {
+            if (this.lhs.isAggregation && this.lhs.list.filter) {
+                const name = this.lhs.list.name;
+                const paramType = schema.inReq[name] || schema.inOpt[name] || schema.out[name];
+                const args = [];
+                if (paramType.elem.isCompound) {
+                    for (let field in paramType.elem.fields)
+                        args.push(new ArgumentDef(null, 'out', field, paramType.elem.fields[field].type, {}, {}));
+                } else {
+                    args.push(new ArgumentDef(null, 'out', 'value', paramType.elem, {}, {}));
+                }
+                const localschema = new ExpressionSignature(null, 'query', null, [], args, {});
+                yield* this.lhs.list.filter.iterateSlots(localschema, prim, makeScope(this));
+            }
+            return;
+        }
+
+        yield [schema, this, prim, scope];
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this boolean expression.
+     *
+     * @alias Ast.BooleanExpression#iterateSlots2
+     * @param {Ast.ExpressionSignature} schema - the signature of the query expression this filter is attached to
+     * @param {Ast.Invocation} prim - the nearest primitive
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2(schema, prim, scope) {
+        if (this.isTrue || this.isFalse)
+            return;
+        if (this.isAnd || this.isOr) {
+            for (let op of this.operands)
+                yield* op.iterateSlots2(schema, prim, scope);
+            return;
+        }
+        if (this.isNot) {
+            yield* this.expr.iterateSlots2(schema, prim, scope);
+            return;
+        }
+        if (this.isCompute) {
+            if (this.lhs.isAggregation && this.lhs.list.filter) {
+                const name = this.lhs.list.name;
+                const paramType = schema.inReq[name] || schema.inOpt[name] || schema.out[name];
+                const args = [];
+                if (paramType.elem.isCompound) {
+                    for (let field in paramType.elem.fields)
+                        args.push(new ArgumentDef(null, 'out', field, paramType.elem.fields[field].type, {}, {}));
+                } else {
+                    args.push(new ArgumentDef(null, 'out', 'value', paramType.elem, {}, {}));
+                }
+                const localschema = new ExpressionSignature(null, 'query', null, [], args, {});
+                yield* this.lhs.list.filter.iterateSlots2(localschema, prim, makeScope(this));
+            }
+            return;
+        }
+        if (this.isVarRef) {
+            //TODO
+            return;
+        }
+        if (this.isExternal) {
+            yield this.selector;
+            yield* iterateSlots2InputParams(this, scope);
+            yield* this.filter.iterateSlots2(this.schema, this, makeScope(this));
+        } else {
+            const arg = schema ? schema.getArgument(this.name) : null;
+            yield* recursiveYieldArraySlots(new FilterSlot(prim, scope, arg, this));
+        }
+    }
+
+    /**
+     * Iterate all primitives (Thingpedia function invocations) in this boolean expression.
+     *
+     * @alias Ast.BooleanExpression#iteratePrimitives
+     * @param {boolean} includeVarRef - whether to include local function calls (VarRef nodes)
+     *                                  in the iteration
+     * @generator
+     * @yields {Ast.Invocation}
+     */
+    *iteratePrimitives() {
+        if (this.isTrue || this.isFalse || this.isAtom)
+            return;
+        if (this.isNot) {
+            yield* this.expr.iteratePrimitives();
+            return;
+        }
+        if (this.isAnd || this.isOr) {
+            for (let op of this.operands)
+                yield* op.iteratePrimitives();
+            return;
+        }
+        if (this.isCompute) {
+            if (this.lhs.isAggregation && this.lhs.list.filter)
+                yield* this.lhs.list.filter.iteratePrimitives();
+            return;
+        }
+        if (this.isVarRef) {
+            // TODO
+            return;
+        }
+
+        assert(this.isExternal);
+        yield ['filter', this];
+        yield* this.filter.iteratePrimitives();
+    }
+}
 BooleanExpression.prototype.isBooleanExpression = true;
 
 /**
