+++ lib/ast/manifest_utils.js
@@ -11,7 +11,7 @@
 
 const Type = require('../type');
 const { Value } = require('./values');
-const { InputParam } = require('./primitive');
+const { InputParam } = require('./expression');
 const { ImportStmt } = require('./program');
 
 function getString(value, fallback='') {
