+++ lib/ast/primitive.js
@@ -12,8 +12,21 @@ const assert = require('assert');
 
 const Node = require('./base');
 const { ExpressionSignature } = require('./function_def');
-const { Invocation, BooleanExpression, ScalarExpression } = require('./expression');
+const { Invocation, Selector, BooleanExpression, ScalarExpression } = require('./expression');
 const { Value } = require('./values');
+const {
+    iterateSlots2InputParams,
+    recursiveYieldArraySlots,
+    makeScope,
+    ArrayIndexSlot,
+    FieldSlot,
+} = require('./slots');
+const { isUnaryStreamToStreamOp,
+        isUnaryTableToTableOp,
+        isUnaryStreamToTableOp,
+        isUnaryTableToStreamOp } = require('../utils');
+const Type = require('../type');
+const Builtin = require('../builtin/defs');
 
 /**
  * The base class of all ThingTalk query expressions.
@@ -35,6 +48,139 @@ class Table extends Node {
 
         assert(schema === null || schema instanceof ExpressionSignature);
         this.schema = schema;
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this table.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.Table#iterateSlots2} instead.
+     */
+    *iterateSlots(scope) {
+        if (this.isVarRef) {
+            for (let in_param of this.in_params)
+                yield [this.schema, in_param, this, scope];
+            return [this, makeScope(this)];
+        } else if (this.isResultRef) {
+            return [this, makeScope(this)];
+        } else if (this.isInvocation) {
+            return yield* this.invocation.iterateSlots(scope);
+        } else if (this.isFilter) {
+            let [prim, newScope] = yield* this.table.iterateSlots(scope);
+            yield* this.filter.iterateSlots(this.table.schema, prim, newScope);
+            return [prim, newScope];
+        } else if (this.isProjection) {
+            let [prim, nestedScope] = yield* this.table.iterateSlots(scope);
+            if (nestedScope === null)
+                return [prim, null];
+            let newScope = {};
+            for (let name of this.args)
+                newScope[name] = nestedScope[name];
+            return [prim, newScope];
+        } else if (isUnaryTableToTableOp(this)) {
+            return yield* this.table.iterateSlots(scope);
+        } else if (isUnaryStreamToTableOp(this)) {
+            return yield* this.stream.iterateSlots(scope);
+        } else if (this.isJoin) {
+            let [, leftScope] = yield* this.lhs.iterateSlots(scope);
+            let [, rightScope] = yield* this.rhs.iterateSlots(scope);
+            if (leftScope === null || rightScope === null)
+                return [null, null];
+            let newScope = {};
+            Object.assign(newScope, leftScope, rightScope);
+            return [null, newScope];
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this table.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2(scope) {
+        if (this.isVarRef) {
+            return yield* iterateSlots2InputParams(this, scope);
+        } else if (this.isResultRef) {
+            const innerScope = makeScope(this);
+            yield* recursiveYieldArraySlots(new FieldSlot(this, innerScope, Type.Number, this, 'result_ref', 'index'));
+            return [this, innerScope];
+        } else if (this.isInvocation) {
+            return yield* this.invocation.iterateSlots2(scope);
+        } else if (this.isFilter) {
+            let [prim, newScope] = yield* this.table.iterateSlots2(scope);
+            yield* this.filter.iterateSlots2(this.table.schema, prim, newScope);
+            return [prim, newScope];
+        } else if (this.isProjection) {
+            let [prim, nestedScope] = yield* this.table.iterateSlots2(scope);
+            if (nestedScope === null)
+                return [prim, null];
+            let newScope = {};
+            for (let name of this.args)
+                newScope[name] = nestedScope[name];
+            return [prim, newScope];
+        } else if (this.isIndex) {
+            const [prim, innerScope] = yield* this.table.iterateSlots2(scope);
+            for (let i = 0; i < this.indices.length; i++)
+                yield* recursiveYieldArraySlots(new ArrayIndexSlot(prim, innerScope, Type.Number, this.indices, 'table.index', i));
+            return [prim, innerScope];
+        } else if (this.isSlice) {
+            const [prim, innerScope] = yield* this.table.iterateSlots2(scope);
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, Type.Number, this, 'slice', 'base'));
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, Type.Number, this, 'slice', 'limit'));
+            return [prim, innerScope];
+        } else if (isUnaryTableToTableOp(this)) {
+            return yield* this.table.iterateSlots2(scope);
+        } else if (isUnaryStreamToTableOp(this)) {
+            return yield* this.stream.iterateSlots2(scope);
+        } else if (this.isJoin) {
+            let [, leftScope] = yield* this.lhs.iterateSlots2(scope);
+            let [, rightScope] = yield* this.rhs.iterateSlots2(scope);
+            if (leftScope === null || rightScope === null)
+                return [null, null];
+            let newScope = {};
+            Object.assign(newScope, leftScope, rightScope);
+            return [null, newScope];
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
+    }
+
+    /**
+     * Iterate all primitives (Thingpedia function invocations) in this table.
+     *
+     * @param {boolean} includeVarRef - whether to include local function calls (VarRef nodes)
+     *                                  in the iteration
+     * @generator
+     * @yields {Ast.Invocation}
+     */
+    *iteratePrimitives(includeVarRef) {
+        if (this.isVarRef) {
+            if (includeVarRef)
+                yield ['query', this];
+        } else if (this.isResultRef) {
+            if (includeVarRef)
+                yield ['query', this];
+        } else if (this.isInvocation) {
+            yield ['query', this.invocation];
+        } else if (this.isFilter) {
+            yield* this.table.iteratePrimitives(includeVarRef);
+            yield* this.filter.iteratePrimitives();
+        } else if (isUnaryTableToTableOp(this)) {
+            yield* this.table.iteratePrimitives(includeVarRef);
+        } else if (isUnaryStreamToTableOp(this)) {
+            yield* this.stream.iteratePrimitives(includeVarRef);
+        } else if (this.isJoin) {
+            yield* this.lhs.iteratePrimitives(includeVarRef);
+            yield* this.rhs.iteratePrimitives(includeVarRef);
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
     }
 }
 Table.prototype.isTable = true;
@@ -598,6 +744,140 @@ class Stream extends Node {
         assert(schema === null || schema instanceof ExpressionSignature);
         this.schema = schema;
     }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this stream.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.Stream#iterateSlots2} instead.
+     */
+    *iterateSlots(scope) {
+        if (this.isVarRef) {
+            for (let in_param of this.in_params)
+                yield [this.schema, in_param, this, scope];
+            return [this, makeScope(this)];
+        } else if (this.isTimer || this.isAtTimer) {
+            // no primitive here
+            return [null, {}];
+        } else if (this.isFilter || this.isEdgeFilter) {
+            let [prim, newScope] = yield* this.stream.iterateSlots(scope);
+            yield* this.filter.iterateSlots(this.stream.schema, prim, newScope);
+            return [prim, newScope];
+        } else if (this.isProjection) {
+            let [prim, nestedScope] = yield* this.stream.iterateSlots(scope);
+            if (nestedScope === null)
+                return [prim, null];
+            let newScope = {};
+            for (let name of this.args)
+                newScope[name] = nestedScope[name];
+            return [prim, newScope];
+        } else if (isUnaryStreamToStreamOp(this)) {
+            return yield* this.stream.iterateSlots(scope);
+        } else if (isUnaryTableToStreamOp(this)) {
+            return yield* this.table.iterateSlots(scope);
+        } else if (this.isJoin) {
+            let [, leftScope] = yield* this.stream.iterateSlots(scope);
+            let [, rightScope] = yield* this.table.iterateSlots(scope);
+            if (leftScope === null || rightScope === null)
+                return [null, null];
+            let newScope = {};
+            Object.assign(newScope, leftScope, rightScope);
+            return [null, newScope];
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this stream.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2(scope) {
+        if (this.isVarRef) {
+            return yield* iterateSlots2InputParams(this, scope);
+        } else if (this.isTimer) {
+            // no primitive here
+            yield* recursiveYieldArraySlots(new FieldSlot(null, scope, Type.Date, this, 'timer', 'base'));
+            yield* recursiveYieldArraySlots(new FieldSlot(null, scope, Type.Measure('ms'), this, 'timer', 'interval'));
+            return [null, {}];
+        } else if (this.isAtTimer) {
+            for (let i = 0; i < this.time.length; i++)
+                yield* recursiveYieldArraySlots(new ArrayIndexSlot(null, scope, Type.Time, this.time, 'attimer.time', i));
+            if (this.expiration_date !== null)
+                yield* recursiveYieldArraySlots(new FieldSlot(null, scope, Type.Date, this, 'attimer', 'expiration_date'));
+            return [null, {}];
+        } else if (this.isWindow || this.isTimeSeries) {
+            const [prim, innerScope] = this.stream.iterateSlots2(scope);
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, this.isWindow ? Type.Number : Type.Date, this, 'history', 'base'));
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, this.isWindow ? Type.Number : Type.Measure('ms'), this, 'history', 'delta'));
+            return [prim, innerScope];
+        } else if (this.isHistory || this.isSequence) {
+            const [prim, innerScope] = this.table.iterateSlots2(scope);
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, this.isSequence ? Type.Number : Type.Date, this, 'history', 'base'));
+            yield* recursiveYieldArraySlots(new FieldSlot(prim, innerScope, this.isSequence ? Type.Number : Type.Measure('ms'), this, 'history', 'delta'));
+            return [prim, innerScope];
+        } else if (this.isFilter || this.isEdgeFilter) {
+            let [prim, newScope] = yield* this.stream.iterateSlots2(scope);
+            yield* this.filter.iterateSlots2(this.stream.schema, prim, newScope);
+            return [prim, newScope];
+        } else if (this.isProjection) {
+            let [prim, nestedScope] = yield* this.stream.iterateSlots2(scope);
+            if (nestedScope === null)
+                return [prim, null];
+            let newScope = {};
+            for (let name of this.args)
+                newScope[name] = nestedScope[name];
+            return [prim, newScope];
+        } else if (isUnaryStreamToStreamOp(this)) {
+            return yield* this.stream.iterateSlots2(scope);
+        } else if (isUnaryTableToStreamOp(this)) {
+            return yield* this.table.iterateSlots2(scope);
+        } else if (this.isJoin) {
+            let [, leftScope] = yield* this.stream.iterateSlots2(scope);
+            let [, rightScope] = yield* this.table.iterateSlots2(scope);
+            if (leftScope === null || rightScope === null)
+                return [null, null];
+            let newScope = {};
+            Object.assign(newScope, leftScope, rightScope);
+            return [null, newScope];
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
+    }
+
+    /**
+     * Iterate all primitives (Thingpedia function invocations) in this stream.
+     *
+     * @param {boolean} includeVarRef - whether to include local function calls (VarRef nodes)
+     *                                  in the iteration
+     * @generator
+     * @yields {Ast.Invocation}
+     */
+    *iteratePrimitives(includeVarRef) {
+        if (this.isVarRef) {
+            if (includeVarRef)
+                yield ['stream', this];
+        } else if (this.isTimer || this.isAtTimer) {
+            // no primitive here
+        } else if (this.isFilter || this.isEdgeFilter) {
+            yield* this.stream.iteratePrimitives(includeVarRef);
+            yield* this.filter.iteratePrimitives();
+        } else if (isUnaryStreamToStreamOp(this)) {
+            yield* this.stream.iteratePrimitives(includeVarRef);
+        } else if (isUnaryTableToStreamOp(this)) {
+            yield* this.table.iteratePrimitives(includeVarRef);
+        } else if (this.isJoin) {
+            yield* this.stream.iteratePrimitives(includeVarRef);
+            yield* this.table.iteratePrimitives(includeVarRef);
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
+    }
 }
 Stream.prototype.isStream = true;
 class VarRefStream extends Stream {
@@ -981,6 +1261,68 @@ class Action extends Node {
          * @type {Ast.ExpressionSignature|null}
          */
         this.schema = schema;
+    }
+
+    /**
+     * Utility function to create a `notify` or `return` action.
+     *
+     * @param {string} [what=notify] - what action to create
+     * @return {Ast.Action} the action node
+     */
+    static notifyAction(what = 'notify') {
+        return new InvocationAction(null,
+            new Invocation(null, Selector.Builtin, what, [], Builtin.Actions[what]),
+            Builtin.Actions[what]);
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this action.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.Action#iterateSlots2} instead.
+     */
+    *iterateSlots(scope) {
+        if (this.isInvocation) {
+            yield* this.invocation.iterateSlots(scope);
+        } else if (this.isVarRef) {
+            for (let in_param of this.in_params)
+                yield [this.schema, in_param, this, scope];
+        }
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this action.
+     *
+     * @param {Object.<string, Ast~SlotScopeItem>} scope - available names for parameter passing
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2(scope) {
+        if (this.isInvocation)
+            yield* this.invocation.iterateSlots2(scope);
+        else if (this.isVarRef)
+            yield* iterateSlots2InputParams(this, scope);
+    }
+
+    /**
+     * Iterate all primitives (Thingpedia function invocations) in this action.
+     *
+     * @param {boolean} includeVarRef - whether to include local function calls (VarRef nodes)
+     *                                  in the iteration
+     * @generator
+     * @yields {Ast.Invocation}
+     */
+    *iteratePrimitives(includeVarRef) {
+        if (this.isVarRef) {
+            if (includeVarRef)
+                yield ['action', this];
+        } else if (this.isInvocation) {
+            yield ['action', this.invocation];
+        } else {
+            throw new TypeError("Can't handle " + this);
+        }
     }
 }
 Action.prototype.isAction = true;
