+++ lib/ast/program.js
@@ -21,6 +21,13 @@ const {
     recursiveYieldArraySlots,
     FieldSlot
 } = require('./slots');
+const { prettyprint,
+        prettyprintExample,
+        prettyprintDataset } = require('../prettyprint');
+const Typechecking = require('../typecheck');
+const Optimizer = require('../optimize');
+const convertToPermissionRule = require('./convert_to_permission_rule');
+const lowerReturn = require('./lower_return');
 
 /**
  * The base class of all AST nodes that represent complete ThingTalk
@@ -511,6 +518,16 @@ class Dataset extends Statement {
         this.annotations = annotations;
     }
 
+    /**
+     * Convert this dataset to prettyprinted ThingTalk code.
+     *
+     * @param {string} [prefix] - prefix each output line with this string (for indentation)
+     * @return {string} the prettyprinted code
+     */
+    prettyprint(prefix = '') {
+        return prettyprintDataset(this, prefix);
+    }
+
     async visit(visitor) {
         await visitor.enter(this);
         if (await visitor.visitDataset(this)) {
@@ -560,6 +577,37 @@ class Input extends Node {
     *iterateSlots2() {
     }
     *iteratePrimitives(includeVarRef) {
+    }
+
+    /**
+     * Convert this ThingTalk input to prettyprinted ThingTalk code.
+     *
+     * @param {string} [prefix] - prefix each output line with this string (for indentation)
+     * @return {string} the prettyprinted code
+     */
+    prettyprint(short) {
+        return prettyprint(this, short);
+    }
+
+    /**
+     * Typecheck this ThingTalk input.
+     *
+     * This is the main API to typecheck a ThingTalk input.
+     *
+     * @param {SchemaRetriever} schemas - schema retriever object to retrieve Thingpedia information
+     * @param {boolean} [getMeta=false] - retreive natural language metadata during typecheck
+     */
+    typecheck(schemas, getMeta = false) {
+        if (this.isBookkeeping)
+            return Typechecking.typeCheckBookkeeping(this.intent).then(() => this);
+        if (this.isProgram)
+            return Typechecking.typeCheckProgram(this, schemas, getMeta).then(() => this);
+        else if (this.isPermissionRule)
+            return Typechecking.typeCheckPermissionRule(this, schemas, getMeta).then(() => this);
+        else if (this.isMeta)
+            return Typechecking.typeCheckMeta(this, schemas, getMeta).then(() => this);
+        else
+            throw new Error('Invalid Input type');
     }
 }
 
@@ -649,6 +697,26 @@ class Program extends Input {
             this.principal !== null ? this.principal.clone() : null,
             this.oninputs.map((o) => o.clone())
         );
+    }
+
+    optimize() {
+        return Optimizer.optimizeProgram(this);
+    }
+
+    /**
+     * Attempt to convert this program to an equivalent permission rule.
+     *
+     * @param {string} principal - the principal to use as source
+     * @param {string|null} contactName - the display value for the principal
+     * @return {Ast.Input.PermissionRule|null} the new permission rule, or `null` if conversion failed
+     * @alias Ast.Input.Program#convertToPermissionRule
+     */
+    convertToPermissionRule(principal, contactName) {
+        return convertToPermissionRule(this, principal, contactName);
+    }
+
+    lowerReturn(messaging) {
+        return lowerReturn(this, messaging);
     }
 }
 Program.prototype.isProgram = true;
@@ -843,6 +911,63 @@ class Example extends Node {
             this.preprocessed.slice(0),
             Object.assign({}, this.annotations)
         );
+    }
+
+    /**
+     * Convert this example to prettyprinted ThingTalk code.
+     *
+     * @param {string} [prefix] - prefix each output line with this string (for indentation)
+     * @return {string} the prettyprinted code
+     */
+    prettyprint(prefix = '') {
+        return prettyprintExample(this, prefix);
+    }
+
+    /**
+     * Typecheck this example.
+     *
+     * This method can be used to typecheck an example is isolation,
+     * outside of a ThingTalk program. This is useful to typecheck a dataset
+     * and discard examples that do not typecheck without failing the whole dataset.
+     *
+     * @param {SchemaRetriever} schemas - schema retriever object to retrieve Thingpedia information
+     * @param {boolean} [getMeta=false] - retreive natural language metadata during typecheck
+     */
+    typecheck(schemas, getMeta = false) {
+        return Typechecking.typeCheckExample(this, schemas, {}, getMeta);
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this example.
+     *
+     * @generator
+     * @yields {Ast~OldSlot}
+     * @deprecated Use {@link Ast.Example#iterateSlots2} instead.
+     */
+    *iterateSlots() {
+        yield* this.value.iterateSlots();
+    }
+
+    /**
+     * Iterate all slots (scalar value nodes) in this example.
+     *
+     * @generator
+     * @yields {Ast~AbstractSlot}
+     */
+    *iterateSlots2() {
+        yield* this.value.iterateSlots2();
+    }
+
+    /**
+     * Iterate all primitives (Thingpedia function invocations) in this example.
+     *
+     * @param {boolean} includeVarRef - whether to include local function calls (VarRef nodes)
+     *                                  in the iteration
+     * @generator
+     * @yields {Ast.Invocation}
+     */
+    *iteratePrimitives(includeVarRef) {
+        yield* this.value.iteratePrimitives(includeVarRef);
     }
 }
 Example.prototype.isExample = true;
