+++ lib/ast/remote_utils.js
@@ -1,0 +1,24 @@
+// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
+//
+// This file is part of ThingTalk
+//
+// Copyright 2015-2020 The Board of Trustees of the Leland Stanford Junior University
+//
+// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
+//
+// See COPYING for details
+"use strict";
+
+function isRemoteReceive(fn) {
+    return (fn.selector.isDevice && fn.selector.kind === 'org.thingpedia.builtin.thingengine.remote' || fn.selector.kind.startsWith('__dyn_')) &&
+        fn.channel === 'receive';
+}
+function isRemoteSend(fn) {
+    return (fn.selector.isDevice && fn.selector.kind === 'org.thingpedia.builtin.thingengine.remote' || fn.selector.kind.startsWith('__dyn_')) &&
+        fn.channel === 'send';
+}
+
+module.exports = {
+    isRemoteSend,
+    isRemoteReceive
+};
