+++ lib/ast/slots.js
@@ -524,8 +524,31 @@ function* recursiveYieldArraySlots(slot) {
         yield* recursiveYieldArraySlots(new ArrayIndexSlot(slot.primitive, slot.scope, type.elem, value.value, slot, i));
 }
 
+/**
+ * Type used by the old slot iteration API.
+ *
+ * This is actually a tuple but jsdoc does not understand tuples.
+ * @typedef Ast~OldSlot
+ * @property {Ast.ExpressionSignature} 0 - the signature of the nearest primitive
+ * @property {Ast.InputParam|Ast.BooleanExpression.Atom} 1 - the holder of the value
+ * @property {Ast.Invocation} 2 - the nearest primitive
+ * @property {Object.<string, Ast~SlotScopeItem>} 3 - available names for parameter passing
+ * @generator
+ * @deprecated Use {@link Ast~AbstractSlot} and the new slot iteration API
+ */
+
+function* iterateSlots2InputParams(prim, scope) {
+    for (let in_param of prim.in_params) {
+        const arg = prim.schema ? prim.schema.getArgument(in_param.name) : null;
+        yield* recursiveYieldArraySlots(new InputParamSlot(prim, scope, arg, in_param));
+    }
+    return [prim, makeScope(prim)];
+}
+
 module.exports = {
     recursiveYieldArraySlots,
+    iterateSlots2InputParams,
+
     makeScope,
     InputParamSlot,
     DeviceAttributeSlot,
